move skipUnless to test_util
